Recognize MBBSEmu settings config file.
Continuation of the previous commit - if MBBSEmu settings configuration
was passed on the command line (default: appsettings.json), use the
given file instead of the default.

I used a workaround found on Stack Overflow to de-staticize the
ServiceResolver class a bit. I made the minimal changes I needed to make
this compile and run.

In the long haul, more core logic can be moved out of ServiceResolver
and Program to improve testability and reusability.