Do not increment UsrPtr.Substt when exiting FSD.
This line of code was preventing the FSD from returning control to the user when exiting the Majormud character editor.

After conversation on Github and Discord, we determined this code was putting the system into a bad state.

We are not aware of any other modules and/or FSDs that require this code, so we feel confident removing it.